requirements.txt
dj-database-url==0.4.2
Django==2.0
django-countries==5.3
django-crispy-forms==1.7.0
django-money==0.13.3
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
gunicorn==19.7.1
numpy==1.14.5
olefile==0.44
pandas==0.22.0
Pillow==4.3.0
psycopg2==2.7.3.2
py-moneyed==0.7.0
python-dateutil==2.7.3
pytz==2017.3
six==1.11.0
xlrd==1.1.0
